backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
langchain-huggingface
youtube-transcript-api
faiss-cpu
tiktoken
sentence-transformers